[PyTorch] Address THD AG review follow-ups
Resolve review comments for PR 2829 by tightening the THD all_gather output shape, renaming the new THD CP helper bindings, removing the unrelated pybind helper extraction, and aligning the FP8 t3hd aux handling with the post-FP8DS code path.

Also clean up the THD CP test skip logic and remove an unnecessary dtype conversion from the THD max-logit mask construction.

+++ transformer_engine/pytorch/cpp_extensions/fused_attn.py
@@ -371,8 +371,8 @@ def fused_attn_fwd(
                         device=max_tensor.device
                     )
                     tq = max_tensor.shape[0]
-                    starts = cu_seqlens_q_padded[:-1].to(device=max_tensor.device, dtype=torch.long)
-                    ends = (starts + actual_seqlens.to(dtype=torch.long)).clamp(max=tq)
+                    starts = cu_seqlens_q_padded[:-1].to(device=max_tensor.device)
+                    ends = (starts + actual_seqlens).clamp(max=tq)
                     delta = torch.zeros(tq + 1, dtype=torch.int32, device=max_tensor.device)
                     updates = torch.ones_like(starts, dtype=torch.int32)
                     delta.scatter_add_(0, starts.clamp(max=tq), updates)
